+++ taos/cinterface.py
@@ -53,6 +53,7 @@ def _load_taos_windows():
         print("unable to load taos client library: %s" % e)
         try:
             from ctypes.util import find_library
+
             ctypes.windll.LoadLibrary(find_library("taos"))
         except Exception as final_err:
             raise InterfaceError("unable to load taos client library: %s" % final_err)
@@ -360,7 +361,13 @@ def taos_query_a(connection, sql, callback, param):
 try:
     async_query_with_reqid_callback_type = CFUNCTYPE(None, c_void_p, c_void_p, c_int)
     _libtaos.taos_query_a_with_reqid.restype = None
-    _libtaos.taos_query_a_with_reqid.argtypes = c_void_p, c_char_p, async_query_with_reqid_callback_type, c_void_p, c_int
+    _libtaos.taos_query_a_with_reqid.argtypes = (
+        c_void_p,
+        c_char_p,
+        async_query_with_reqid_callback_type,
+        c_void_p,
+        c_int,
+    )
 except Exception as err:
     _UNSUPPORTED["taos_query_a_with_reqid"] = err
 
@@ -378,8 +385,8 @@ def taos_query_a_with_reqid(connection, sql, callback, param, req_id):
     """
     _check_if_supported()
     _libtaos.taos_query_a_with_reqid(
-        connection, c_char_p(sql.encode("utf-8")),
-        async_query_with_reqid_callback_type(callback), param, req_id)
+        connection, c_char_p(sql.encode("utf-8")), async_query_with_reqid_callback_type(callback), param, req_id
+    )
 
 
 async_fetch_rows_callback_type = CFUNCTYPE(None, c_void_p, c_void_p, c_int)
@@ -1029,10 +1036,10 @@ except Exception as err:
 
 
 def taos_schemaless_insert(
-        connection: c_void_p,
-        lines: Union[List[str], Tuple[str]],
-        protocol: SmlProtocol,
-        precision: SmlPrecision,
+    connection: c_void_p,
+    lines: Union[List[str], Tuple[str]],
+    protocol: SmlProtocol,
+    precision: SmlPrecision,
 ):
     _check_if_supported()
     num_of_lines = len(lines)
@@ -1077,11 +1084,11 @@ except Exception as err:
 
 
 def taos_schemaless_insert_ttl(
-        connection: c_void_p,
-        lines: Union[List[str], Tuple[str]],
-        protocol: SmlProtocol,
-        precision: SmlPrecision,
-        ttl: int,
+    connection: c_void_p,
+    lines: Union[List[str], Tuple[str]],
+    protocol: SmlProtocol,
+    precision: SmlPrecision,
+    ttl: int,
 ):
     _check_if_supported()
     num_of_lines = len(lines)
@@ -1128,12 +1135,12 @@ except Exception as err:
 
 
 def taos_schemaless_insert_ttl_with_reqid(
-        connection: c_void_p,
-        lines: Union[List[str], Tuple[str]],
-        protocol: SmlProtocol,
-        precision: SmlPrecision,
-        ttl: int,
-        req_id: int,
+    connection: c_void_p,
+    lines: Union[List[str], Tuple[str]],
+    protocol: SmlProtocol,
+    precision: SmlPrecision,
+    ttl: int,
+    req_id: int,
 ):
     _check_if_supported()
     num_of_lines = len(lines)
@@ -1180,7 +1187,8 @@ def taos_schemaless_insert_with_reqid(connection, lines, protocol, precision, re
     lines_type = ctypes.c_char_p * num_of_lines
     p_lines = lines_type(*lines)
     res = c_void_p(
-        _libtaos.taos_schemaless_insert_with_reqid(connection, p_lines, num_of_lines, protocol, precision, req_id))
+        _libtaos.taos_schemaless_insert_with_reqid(connection, p_lines, num_of_lines, protocol, precision, req_id)
+    )
     errno = taos_errno(res)
     affected_rows = taos_affected_rows(res)
     if errno != 0:
@@ -1209,10 +1217,10 @@ except Exception as err:
 
 
 def taos_schemaless_insert_raw(
-        connection: c_void_p,
-        lines_raw: str,
-        protocol: SmlProtocol,
-        precision: SmlPrecision,
+    connection: c_void_p,
+    lines_raw: str,
+    protocol: SmlProtocol,
+    precision: SmlPrecision,
 ) -> int:
     _check_if_supported()
     length = len(lines_raw)
@@ -1220,16 +1228,7 @@ def taos_schemaless_insert_raw(
     type_p_int = POINTER(c_int)
     total_rows = type_p_int(c_int(0))
 
-    res = c_void_p(
-        _libtaos.taos_schemaless_insert_raw(
-            connection,
-            lines_raw,
-            length,
-            total_rows,
-            protocol,
-            precision
-        )
-    )
+    res = c_void_p(_libtaos.taos_schemaless_insert_raw(connection, lines_raw, length, total_rows, protocol, precision))
 
     errno = taos_errno(res)
     affected_rows = taos_affected_rows(res)
@@ -1265,11 +1264,11 @@ except Exception as err:
 
 
 def taos_schemaless_insert_raw_with_reqid(
-        connection: c_void_p,
-        lines_raw: str,
-        protocol: SmlProtocol,
-        precision: SmlPrecision,
-        req_id: int,
+    connection: c_void_p,
+    lines_raw: str,
+    protocol: SmlProtocol,
+    precision: SmlPrecision,
+    req_id: int,
 ) -> int:
     _check_if_supported()
     length = len(lines_raw)
@@ -1323,11 +1322,11 @@ except Exception as err:
 
 
 def taos_schemaless_insert_raw_ttl(
-        connection: c_void_p,
-        lines_raw: str,
-        protocol: SmlProtocol,
-        precision: SmlPrecision,
-        ttl: int,
+    connection: c_void_p,
+    lines_raw: str,
+    protocol: SmlProtocol,
+    precision: SmlPrecision,
+    ttl: int,
 ) -> int:
     _check_if_supported()
     length = len(lines_raw)
@@ -1382,12 +1381,12 @@ except Exception as err:
 
 
 def taos_schemaless_insert_raw_ttl_with_reqid(
-        connection: c_void_p,
-        lines_raw: str,
-        protocol: SmlProtocol,
-        precision: SmlPrecision,
-        ttl: int,
-        req_id: int,
+    connection: c_void_p,
+    lines_raw: str,
+    protocol: SmlProtocol,
+    precision: SmlPrecision,
+    ttl: int,
+    req_id: int,
 ) -> int:
     _check_if_supported()
     length = len(lines_raw)
@@ -1450,9 +1449,9 @@ def tmq_conf_set(conf, key, value):
     if not isinstance(value, str):
         raise TmqError(msg=f"fail to execute tmq_conf_set({key},{value}), {value} is not string type")
 
-    res = _libtaos.tmq_conf_set(conf, ctypes.c_char_p(key.encode("utf-8")), ctypes.c_char_p(value.encode("utf-8")))
-    if res != 0:
-        raise TmqError(msg=f"fail to execute tmq_conf_set({key},{value}), code={res}", errno=res)
+    tmq_res = _libtaos.tmq_conf_set(conf, ctypes.c_char_p(key.encode("utf-8")), ctypes.c_char_p(value.encode("utf-8")))
+    if tmq_res != 0:
+        raise TmqError(msg=f"fail to execute tmq_conf_set({key},{value}), code={tmq_res}", errno=tmq_res)
 
 
 try:
@@ -1496,8 +1495,21 @@ def tmq_consumer_new(conf, errstrlen=0):
     buf = ctypes.create_string_buffer(errstrlen)
     tmq = cast(_libtaos.tmq_consumer_new(conf, buf, errstrlen), c_void_p)
     if tmq.value is None:
-        raise TmqError("failed on tmq_consumer_new")
+        raise TmqError("failed on tmq_consumer_new()")
     return tmq
+
+
+try:
+    _libtaos.tmq_err2str.restype = c_char_p
+    _libtaos.tmq_err2str.argtypes = (c_int,)
+except Exception as err:
+    _UNSUPPORTED["tmq_err2str"] = err
+
+
+def tmq_err2str(errno):
+    # type (c_int) -> c_char_p
+    _check_if_supported()
+    return c_char_p(_libtaos.tmq_err2str(errno))
 
 
 try:
@@ -1580,7 +1592,7 @@ def tmq_subscribe(tmq, list):
     _check_if_supported()
     res = _libtaos.tmq_subscribe(tmq, list)
     if res != 0:
-        raise TmqError(msg="failed on tmq_subscribe()", errno=res)
+        raise TmqError(msg=f"failed on tmq_subscribe(), errno={res:X}, errmsg={tmq_err2str(res)}", errno=res)
 
 
 try:
@@ -1599,7 +1611,7 @@ def tmq_unsubscribe(tmq):
         # tmq_unsubscribe will subscribe a empty topic list to clear the consumer task.
         return
     if res != 0:
-        raise TmqError(msg=f"failed on tmq_unsubscribe() {res}", errno=res)
+        raise TmqError(msg=f"failed on tmq_unsubscribe(), errno={res:X}, errmsg={tmq_err2str(res)}", errno=res)
 
 
 try:
@@ -1615,7 +1627,7 @@ def tmq_subscription(tmq):
     topics = tmq_list_new()
     res = _libtaos.tmq_subscription(tmq, byref(topics))
     if res != 0:
-        raise TmqError(msg="failed on tmq_subscription()", errno=res)
+        raise TmqError(msg=f"failed on tmq_subscription(), errno={res:X}, errmsg={tmq_err2str(res)}", errno=res)
     res = tmq_list_to_c_array(topics)
     tmq_list_destroy(topics)
     return res
@@ -1646,7 +1658,7 @@ def tmq_consumer_close(tmq):
     _check_if_supported()
     res = _libtaos.tmq_consumer_close(tmq)
     if res != 0:
-        raise TmqError(msg="failed on tmq_consumer_close()", errno=res)
+        raise TmqError(msg=f"failed on tmq_consumer_close(), errno={res:X}, errmsg={tmq_err2str(res)}", errno=res)
 
 
 try:
@@ -1661,7 +1673,7 @@ def tmq_commit_sync(tmq, offset):
     _check_if_supported()
     res = _libtaos.tmq_commit_sync(tmq, offset)
     if res != 0:
-        raise TmqError(msg="failed on tmq_commit_sync()", errno=res)
+        raise TmqError(msg=f"failed on tmq_commit_sync(), errno={res:X}, errmsg={tmq_err2str(res)}", errno=res)
 
 
 try:
@@ -1745,8 +1757,9 @@ def taos_get_table_vgId(conn, db, table):
     # type: (c_void_p, str, str) -> int
     _check_if_supported()
     vg_id = c_int()
-    code = _libtaos.taos_get_table_vgId(conn, c_char_p(db.encode('utf-8')), c_char_p(table.encode('utf-8')),
-                                        ctypes.byref(vg_id))
+    code = _libtaos.taos_get_table_vgId(
+        conn, c_char_p(db.encode("utf-8")), c_char_p(table.encode("utf-8")), ctypes.byref(vg_id)
+    )
     if code != 0:
         raise InternalError(taos_errstr(c_void_p(None)))
     return vg_id.value
@@ -1762,6 +1775,125 @@ except Exception as err:
 def tmq_get_res_type(message):
     # type: (c_void_p) -> int
     return _libtaos.tmq_get_res_type(message)
+
+
+class TmqTopicAssignment(Structure):
+    _fields_ = [
+        ("_vg_id", c_int32),
+        ("_current_offset", c_int64),
+        ("_begin", c_int64),
+        ("_end", c_int64),
+    ]
+
+    @property
+    def vg_id(self):
+        return self._vg_id
+
+    @property
+    def current_offset(self):
+        return self._current_offset
+
+    @property
+    def begin(self):
+        return self._begin
+
+    @property
+    def end(self):
+        return self._end
+
+    def __str__(self):
+        return "vg_id: %s, current_offset: %s, begin: %s, end: %s" % (
+            self.vg_id, self.current_offset, self.begin, self.end
+        )
+
+
+class TmqTopicAssignments(Structure):
+
+    def __init__(self, assignments, count):
+        self._assignments = []
+        if isinstance(assignments, c_void_p):
+            self._assignments = cast(assignments, POINTER(TmqTopicAssignment))
+        if isinstance(assignments, POINTER(TmqTopicAssignment)):
+            self._assignments = assignments
+        self._count = count
+        self._iter = 0
+
+    def as_ptr(self):
+        return self._assignments
+
+    @property
+    def count(self):
+        return self._count
+
+    @property
+    def assignments(self):
+        return self._assignments
+
+    def next(self):
+        return self._next()
+
+    def _next(self):
+        if self._iter < self.count:
+            assignment = self._assignments[self._iter]
+            self._iter += 1
+            return assignment
+        else:
+            raise StopIteration
+
+    def __next__(self):
+        return self._next()
+
+    def __getitem__(self, item):
+        return self._assignments[item]
+
+    def __iter__(self):
+        self._iter = 0
+        return self
+
+    def __len__(self):
+        return self._count
+
+
+try:
+    _libtaos.tmq_get_topic_assignment.argstype = (c_void_p, c_char_p, c_void_p, POINTER(c_int))
+    _libtaos.tmq_get_topic_assignment.restype = c_int
+except Exception as err:
+    _UNSUPPORTED["tmq_get_topic_assignment"] = err
+
+
+def tmq_get_topic_assignment(tmq, topic_name):
+    # type: (c_void_p, str) -> List[()]
+
+    _check_if_supported()
+    num_of_assignment = c_int()
+    assignment = c_void_p()
+    assignments = []
+    code = _libtaos.tmq_get_topic_assignment(tmq, c_char_p(topic_name.encode('utf-8')), byref(assignment),
+                                             ctypes.byref(num_of_assignment))
+    if code != 0:
+        raise TmqError(msg="failed on tmq_get_topic_assignment()", errno=code)
+
+    tmq_assignments = TmqTopicAssignments(assignment, num_of_assignment.value)
+
+    for tmq_assignment in tmq_assignments:
+        assignments.append(
+            (tmq_assignment.vg_id, tmq_assignment.current_offset, tmq_assignment.begin, tmq_assignment.end))
+
+    _libtaos.tmq_free_assignment(assignment)
+    return assignments
+
+
+try:
+    _libtaos.tmq_offset_seek.argstype = (c_void_p, c_char_p, c_int32, c_int64)
+    _libtaos.tmq_offset_seek.restype = c_int
+except Exception as err:
+    _UNSUPPORTED["tmq_offset_seek"] = err
+
+
+def tmq_offset_seek(tmq, topic_name, vgroup_id, offset):
+    code = _libtaos.tmq_offset_seek(tmq, c_char_p(topic_name.encode('utf-8')), c_int32(vgroup_id), c_int64(offset))
+    if code != 0:
+        raise TmqError(msg="failed on tmq_offset_seek()", errno=code)
 
 
 class CTaosInterface(object):